+++ src/main/kotlin/com/projects/inGameMarketplace/inventoryService/InventoryRepository.kt
@@ -47,7 +47,7 @@ class InventoryRepository {
         while(resultSet.next()) {
             val name = resultSet.getString("item_name")
             val image = resultSet.getString("image_url")
-            val price = resultSet.getDouble("price")
+            val price = resultSet.getInt("price")
             val amount = resultSet.getInt("amount")
 
             val item = Item(name, image, price)
